Move NcdDot next to NdcRange component
