+++ src/lib/charts/components/NdcDot.svelte
